+++ src/main/java/com/iti/tictactoe/FullMenuController.java
@@ -1,5 +1,6 @@
 package com.iti.tictactoe;
 
+import com.iti.tictactoe.auth.LoginScreen;
 import com.iti.tictactoe.muliplayerOffline.models.UiUtils;
 import com.iti.tictactoe.navigation.NavigationController;
 import javafx.fxml.FXML;
